+++ llm_gent/core/tools/builtin/__init__.py
@@ -1,5 +1,6 @@
 """Built-in tools for common operations."""
 
+from .brave_search import BraveSearchBackend
 from .complete import CompleteTaskTool
 from .file import FileReadTool, FileWriteTool
 from .http import HTTPFetchTool
@@ -10,6 +11,7 @@ from .web_search import WebSearchTool
 
 
 __all__ = [
+    "BraveSearchBackend",
     "CompleteTaskTool",
     "FileReadTool",
     "FileWriteTool",